NumPy
NumPy is one of the most important libraries to data analysis. It is a numeric computing library that does mathematical operations efficiently. Since computers can only process 1's and 0's and not decimal numbers processing code can take up lots of space. Hence, using NumPy allows you to control size in terms of bits. In NumPy there are  a few objects [ OBJECT- a collection of data and or variables in methods]. You can store data in lists or in arrays. Arrays are a  container which can hold a fix number of items and these items should be of the same type.

In my practice I learned how to create arrays, index arrays, slice arrays, and determine the size, shape, and dimension of arrays. This was quite fun and I hope you enjoy reading my code.